Restore xml_security.rb.
libxmlsec1 is strict with trust chain verification. Until it's possible
to easily trust a cert, this causes a problem for integration tests with
service providers where the CA is not easily accessible (eg if the SP CA
is firewalled off from the IdP server). Reverting to xml_security.rb
allows us to ignore certificate chain verification.