Ensure NameID formats actually match the spec (#36)
Resolves LG-4273

Section 8.3 of the SAML 2.0 spec specifies that the email address NameID
format should be versioned at 1.1, but earlier work has modified this
library such that the metadata generator specifies it at 2.0. This makes
things consistent with the spec by checking an individual format for the
appropriate version number before generating it.